backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
httpx
pydantic-settings
python-dotenv
email-validator
